code optimization and readability improvment
