+++ 02_knight_and_princess/anton_potapov/graph.json
@@ -1,5 +1,3 @@
-Получаемый json файл: <br />
-```javascript
 {
 	"vertices": [
 		{
@@ -133,4 +131,3 @@
 		}
 	]
 }
-```
