+++ 02_knight_and_princess/anton_potapov/properties.hpp
