--- task_01/anton_potapov/README.md
@@ -1,2 +1,0 @@
-Антон Потапов <br />
-![Screenshot](screenshot.png)
