--- task_01/anton_potapov/hello_world.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-
-int main() {
-  std::cout << "Hello World!" << std::endl << "Антон Потапов" << std::endl;
-  return 0;
-}
