--- task_01/nikolay_shapurov/README.md
@@ -1,3 +1,0 @@
-Шапуров Николай
-
-<img src="picture.png" width="1200" height="800">
