--- task_01/nikolay_shapurov/hello_world.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-
-int main() {
-  std::cout << "Hello World!" << std::endl << "Николай Шапуров" << std::endl;
-  return 0;
-}
