--- task_02/anton_potapov/.gitignore
@@ -1,2 +1,0 @@
-a.out
-graph.json
